refactor: inline RevisionService into AbstractListService
