+++ src/Services/List/PluginListService.php
@@ -5,14 +5,14 @@ declare(strict_types=1);
 namespace App\Services\List;
 
 use App\ResourceType;
-use App\Services\Interfaces\RevisionServiceInterface;
 use App\Services\Interfaces\SubversionServiceInterface;
 use App\Services\Metadata\PluginMetadataService;
+use Doctrine\ORM\EntityManagerInterface;
 
-readonly class PluginListService extends AbstractListService
+class PluginListService extends AbstractListService
 {
-    public function __construct(SubversionServiceInterface $svn, PluginMetadataService $meta, RevisionServiceInterface $revisions)
+    public function __construct(SubversionServiceInterface $svn, PluginMetadataService $meta, EntityManagerInterface $em)
     {
-        parent::__construct($svn, $meta, $revisions, ResourceType::Plugin);
+        parent::__construct($svn, $meta, $em, ResourceType::Plugin);
     }
 }
