+++ src/Services/List/ThemeListService.php
@@ -5,14 +5,14 @@ declare(strict_types=1);
 namespace App\Services\List;
 
 use App\ResourceType;
-use App\Services\Interfaces\RevisionServiceInterface;
 use App\Services\Interfaces\SubversionServiceInterface;
 use App\Services\Metadata\ThemeMetadataService;
+use Doctrine\ORM\EntityManagerInterface;
 
-readonly class ThemeListService extends AbstractListService
+class ThemeListService extends AbstractListService
 {
-    public function __construct(SubversionServiceInterface $svn, ThemeMetadataService $meta, RevisionServiceInterface $revisions)
+    public function __construct(SubversionServiceInterface $svn, ThemeMetadataService $meta, EntityManagerInterface $em)
     {
-        parent::__construct($svn, $meta, $revisions, ResourceType::Theme);
+        parent::__construct($svn, $meta, $em, ResourceType::Theme);
     }
 }
